+++ R/plot.p_significance.R
@@ -1,7 +1,7 @@
 #' @importFrom insight clean_parameters
 #' @importFrom dplyr group_by mutate ungroup select one_of n
 #' @export
-data_plot.p_direction <- function(x, data = NULL, show_intercept = FALSE, ...){
+data_plot.p_significance <- function(x, data = NULL, grid = TRUE, show_intercept = FALSE, ...){
   if (is.null(data)) {
     data <- .retrieve_data(x)
   }
@@ -29,13 +29,13 @@ data_plot.p_direction <- function(x, data = NULL, show_intercept = FALSE, ...){
         dataplot <- rbind(
           dataplot,
           cbind(
-            .compute_densities_pd(data[[i]], name = i),
+            .compute_densities_ps(data[[i]], name = i, threshold = attr(x, "threshold")),
             "Effects" = params$Effects[params$Parameter == i],
             "Component" = params$Component[params$Parameter == i]
           )
         )
       } else {
-        dataplot <- rbind(dataplot, .compute_densities_pd(data[[i]], name = i))
+        dataplot <- rbind(dataplot, .compute_densities_ps(data[[i]], name = i, threshold = attr(x, "threshold")))
       }
     }
 
@@ -84,11 +84,11 @@ data_plot.p_direction <- function(x, data = NULL, show_intercept = FALSE, ...){
   dataplot <- .fix_facet_names(dataplot)
 
   attr(dataplot, "info") <- list("xlab" = "Possible parameter values",
-                                  "ylab" = ylab,
-                                  "legend_fill" = "Effect direction",
-                                  "title" = "Probability of Direction")
+                                 "ylab" = ylab,
+                                 "legend_fill" = "Probability",
+                                 "title" = "Practical Significance")
 
-  class(dataplot) <- c("data_plot", "see_p_direction", class(dataplot))
+  class(dataplot) <- c("data_plot", "see_p_significance", class(dataplot))
   dataplot
 }
 
@@ -98,12 +98,20 @@ data_plot.p_direction <- function(x, data = NULL, show_intercept = FALSE, ...){
 #' @importFrom stats density
 #' @importFrom dplyr mutate
 #' @keywords internal
-.compute_densities_pd <- function(x, name = "Y"){
+.compute_densities_ps <- function(x, name = "Y", threshold = 0) {
   out <- x %>%
     stats::density() %>%
-    .as.data.frame_density() %>%
-    dplyr::mutate(fill = ifelse(.data$x < 0, "Negative", "Positive")) %>%
-    dplyr::mutate(height = .data$y, y = name)
+    .as.data.frame_density()
+
+  fifty_cents <- sum(out$y[out$x > threshold]) > (sum(out$y) / 2)
+
+  out$fill <- "Less Probable"
+  out$fill[abs(out$x) < threshold] <- "ROPE"
+  out$fill[(out$x > threshold)] <- ifelse(fifty_cents, "Significant", "Less Probable")
+  out$fill[out$x < (-1 * threshold)] <- ifelse(fifty_cents, "Less Probable", "Significant")
+
+  out$height <- out$y
+  out$y <- name
 
   # normalize
   out$height <- as.vector((out$height - min(out$height, na.rm = TRUE)) / diff(range(out$height, na.rm = TRUE), na.rm = TRUE))
@@ -117,7 +125,7 @@ data_plot.p_direction <- function(x, data = NULL, show_intercept = FALSE, ...){
 #' @importFrom ggridges geom_ridgeline_gradient
 #' @rdname data_plot
 #' @export
-plot.see_p_direction <- function(x, data = NULL, show_intercept = FALSE, priors = FALSE, priors_alpha = .4, n_columns = 1, ...) {
+plot.see_p_significance <- function(x, data = NULL, show_intercept = FALSE, priors = FALSE, priors_alpha = .4, n_columns = 1, ...) {
   # save model for later use
   model <- .retrieve_data(x)
 
@@ -148,14 +156,21 @@ plot.see_p_direction <- function(x, data = NULL, show_intercept = FALSE, priors 
 
   # add prior layer
   if (priors) {
-    p <- p + .add_prior_layer_ridgeline(
-      model,
-      show_intercept = show_intercept,
-      priors_alpha = priors_alpha
-    )
+    p <- p +
+      .add_prior_layer_ridgeline(
+        model,
+        show_intercept = show_intercept,
+        priors_alpha = priors_alpha
+      ) +
+      scale_fill_manual(values = c("white", "#FFFC00", "#cd201f", "#0077B5"))
+  } else {
+    p <- p +
+      scale_fill_manual(values = c("white", "#cd201f", "#0077B5"))
   }
 
-  p <- p + geom_vline(aes(xintercept = 0), linetype = "dotted")
+  p <- p +
+    geom_vline(aes(xintercept = 0), linetype = "dotted") +
+    guides(fill = FALSE, color = FALSE, group = FALSE)
 
 
   if (length(unique(x$y)) == 1 && is.numeric(x$y)) {