+++ R/plot.point_estimates.R
@@ -10,9 +10,17 @@ data_plot.point_estimate <- function(x, data = NULL, ...) {
       stop("Package 'emmeans' required for this function to work. Please install it.", call. = FALSE)
     }
     data <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(data, names = FALSE)))
+  } else if (inherits(data, c("stanreg", "brmsfit"))) {
+    params <- insight::clean_parameters(data)
+    data <- insight::get_parameters(data, effects = "all", component = "all")
   } else {
     data <- as.data.frame(data)
   }
+
+  data <- tryCatch(
+    {data[, x$Parameter]},
+    error = function(e) {data}
+  )
 
   centrality <- tolower(attr(x, "centrality", exact = TRUE))
   if (is.null(centrality)) centrality <- "all"
@@ -43,6 +51,8 @@ data_plot.point_estimate <- function(x, data = NULL, ...) {
   })
 
   class(dataplot) <- c("data_plot", "see_point_estimate", class(dataplot))
+  names(dataplot) <- colnames(data)
+
   dataplot
 }
 
@@ -54,13 +64,12 @@ data_plot.map_estimate <- data_plot.point_estimate
 
 # Plot --------------------------------------------------------------------
 #' @importFrom rlang .data
-#' @param show_labels Logical, if \code{TRUE}, the text labels for the point
-#'   estimates (i.e. \emph{"Mean"}, \emph{"Median"} and/or \emph{"MAP"}) are shown.
-#'   You may set \code{show_labels = FALSE} in case of overlapping labels, and
-#'   add your own legend or footnote to the plot.
 #' @rdname data_plot
 #' @export
-plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 3.5, panel = TRUE, show_labels = TRUE, show_intercept = FALSE, ...){
+plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 3.5, panel = TRUE, show_labels = TRUE, show_intercept = FALSE, priors = FALSE, priors_alpha = .4, ...){
+  # save model for later use
+  model <- .retrieve_data(x)
+
   if (!"data_plot" %in% class(x)) {
     x <- data_plot(x, data = data)
   }
@@ -75,11 +84,11 @@ plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 
     max_y <- max(i$y)
 
     if ("group" %in% colnames(i))
-      x_lab <- unique(i$group)
+      x_lab <- .clean_parameter_names(unique(i$group))
     else
       x_lab <- "Parameter Value"
 
-    if (!show_intercept && x_lab %in% c("Intercept", "(Intercept)", "b_Intercept")) return(NULL)
+    if (!show_intercept && .has_intercept(x_lab)) return(NULL)
 
     label_mean_x <- mean_x
     label_mean_y <- max_y * 1.05
@@ -88,13 +97,29 @@ plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 
     label_map_x <- map_x
     label_map_y <- max_y * 1.05
 
-    p_object <- ggplot(i, aes(x = .data$x, y = .data$y, group = .data$group)) +
-      geom_ribbon(aes(ymin = 0, ymax = .data$y), fill = "#FFC107")
+    p_object <- ggplot(i, aes(x = .data$x, y = .data$y, group = .data$group))
+
+    # add prior layer
+    if (priors) {
+      p_object <- p_object + .add_prior_layer_ribbon(
+        model,
+        parameter = x_lab,
+        show_intercept = show_intercept,
+        priors_alpha = priors_alpha,
+        fill_color = "#FF9800"
+      )
+      posterior_alpha <- .7
+    } else {
+      posterior_alpha <- 1
+    }
+
+    p_object <- p_object +
+      geom_ribbon(aes(ymin = 0, ymax = .data$y), fill = "#FFC107", alpha = posterior_alpha)
 
     if (!is.null(mean_x) && !is.null(mean_y)) {
       p_object <- p_object +
-        geom_segment(x = mean_x, xend = mean_x, y = 0, yend = mean_y, color = "#E91E63", size = 1) +
-        geom_point(x = mean_x, y = mean_y, color = "#E91E63", size = point_size)
+        geom_segment(x = mean_x, xend = mean_x, y = 0, yend = mean_y, color = "#E91E63", size = 1, alpha = posterior_alpha) +
+        geom_point(x = mean_x, y = mean_y, color = "#E91E63", size = point_size, alpha = posterior_alpha)
       if (show_labels) {
         p_object <- p_object +
           geom_text(x = label_mean_x, y = label_mean_y, label = "Mean", color = "#E91E63", size = text_size)
@@ -103,8 +128,8 @@ plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 
 
     if (!is.null(median_x) && !is.null(median_y)) {
       p_object <- p_object +
-        geom_segment(x = median_x, xend = median_x, y = 0, yend = median_y, color = "#2196F3", size = 1) +
-        geom_point(x = median_x, y = median_y, color = "#2196F3", size = point_size)
+        geom_segment(x = median_x, xend = median_x, y = 0, yend = median_y, color = "#2196F3", size = 1, alpha = posterior_alpha) +
+        geom_point(x = median_x, y = median_y, color = "#2196F3", size = point_size, alpha = posterior_alpha)
       if (show_labels) {
         p_object <- p_object +
           geom_text(x = label_median_x, y = label_median_y, label = "Median", color = "#2196F3", size = text_size)
@@ -113,8 +138,8 @@ plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 
 
     if (!is.null(map_x) && !is.null(map_y)) {
       p_object <- p_object +
-        geom_segment(x = map_x, xend = map_x, y = 0, yend = map_y, color = "#4CAF50", size = 1) +
-        geom_point(x = map_x, y = map_y, color = "#4CAF50", size = point_size)
+        geom_segment(x = map_x, xend = map_x, y = 0, yend = map_y, color = "#4CAF50", size = 1, alpha = posterior_alpha) +
+        geom_point(x = map_x, y = map_y, color = "#4CAF50", size = point_size, alpha = posterior_alpha)
       if (show_labels) {
         p_object <- p_object +
           geom_text(x = label_map_x, y = label_map_y, label = "MAP", color = "#4CAF50", size = text_size)
@@ -124,7 +149,6 @@ plot.see_point_estimate <- function(x, data = NULL, point_size = 2, text_size = 
     p_object <- p_object +
       geom_vline(xintercept = 0, linetype = "dotted") +
       scale_y_continuous(expand = c(0, 0), limits = c(0, max_y * 1.15)) +
-      see::theme_lucid() +
       labs(title = "Bayesian Point Estimates", x = x_lab, y = "Probability Density")
 
     p_object