+++ R/plot.rope.R
@@ -1,15 +1,20 @@
 #' @importFrom dplyr group_by mutate ungroup select one_of n
 #' @export
-data_plot.rope <- function(x, data = NULL, ...){
+data_plot.rope <- function(x, data = NULL, show_intercept = FALSE, ...){
   if (is.null(data)) {
     data <- .retrieve_data(x)
   }
+
+  params <- NULL
 
   if (inherits(data, "emmGrid")) {
     if (!requireNamespace("emmeans", quietly = TRUE)) {
       stop("Package 'emmeans' required for this function to work. Please install it.", call. = FALSE)
     }
     data <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(data, names = FALSE)))
+  } else if (inherits(data, c("stanreg", "brmsfit"))) {
+    params <- insight::clean_parameters(data)
+    data <- as.data.frame(data)
   } else {
     data <- as.data.frame(data)
   }
@@ -25,11 +30,23 @@ data_plot.rope <- function(x, data = NULL, ...){
   }
 
   # Extract data HDI
-  dataplot <- .data_plot_hdi(hdi, data)
+  dataplot <- .data_plot_hdi(hdi, data, parms = params)
   rope_range <- unique(c(x$ROPE_low, x$ROPE_high))
   if (length(rope_range) != 2) {
     stop("Only one ROPE range accepted.")
   }
+
+  groups <- unique(dataplot$y)
+  if (!show_intercept) {
+    dataplot <- .remove_intercept(dataplot, column = "y", show_intercept = show_intercept)
+    groups <- unique(setdiff(groups, .intercepts()))
+  }
+
+  if (length(groups) == 1) {
+    dataplot$y <- 0
+  }
+
+  dataplot <- .fix_facet_names(dataplot)
 
   dataplot$xmin <- rope_range[1]
   dataplot$xmax <- rope_range[2]
@@ -50,15 +67,18 @@ data_plot.rope <- function(x, data = NULL, ...){
 # Plot --------------------------------------------------------------------
 #' @rdname data_plot
 #' @importFrom rlang .data
-#' @param rope_alpha Transparency level of ROPE ribbon.
-#' @param rope_color Color of ROPE ribbon.
 #' @export
-plot.see_rope <- function(x, data = NULL, rope_alpha = 0.5, rope_color = "cadetblue", show_intercept = FALSE, ...) {
+plot.see_rope <- function(x, data = NULL, rope_alpha = 0.5, rope_color = "cadetblue", show_intercept = FALSE, n_columns = 1, ...) {
   if (!"data_plot" %in% class(x)) {
-    x <- data_plot(x, data = data)
+    x <- data_plot(x, data = data, show_intercept = show_intercept)
   }
 
-  x <- .remove_intercept(x, column = "y", show_intercept = show_intercept)
+  # check if we have multiple panels
+  if ((!"Effects" %in% names(x) || length(unique(x$Effects)) <= 1) &&
+      (!"Component" %in% names(x) || length(unique(x$Component)) <= 1)) n_columns <- NULL
+
+  # get labels
+  labels <- .clean_parameter_names(x$y, grid = !is.null(n_columns))
 
   p <- x %>%
     as.data.frame() %>%
@@ -80,6 +100,23 @@ plot.see_rope <- function(x, data = NULL, rope_alpha = 0.5, rope_color = "cadetb
       alpha = rope_alpha
     ) +
     add_plot_attributes(x)
+
+  if (length(unique(x$y)) == 1 && is.numeric(x$y)) {
+    p <- p + scale_y_continuous(breaks = NULL, labels = NULL)
+  } else {
+    p <- p + scale_y_discrete(labels = labels)
+  }
+
+  if (!is.null(n_columns)) {
+    if ("Component" %in% names(x) && "Effects" %in% names(x)) {
+      p <- p + facet_wrap(~ Effects + Component, scales = "free", ncol = n_columns)
+    } else if ("Effects" %in% names(x)) {
+      p <- p + facet_wrap(~ Effects, scales = "free", ncol = n_columns)
+    } else if ("Component" %in% names(x)) {
+      p <- p + facet_wrap(~ Component, scales = "free", ncol = n_columns)
+    }
+  }
+
   p
 }